Added a new version of an N-gram dictionary.
The new version is a tree structure that is capable of holding n-grams
of all sizes simultaneously.

Each node of the tree contains the a word, likelihood of that word
terminating an n-gram sequence (where the sequence is represented by the
path taken down the tree), and references to more nodes for longer
n-grams.